Update the background color of uneditable parts
Make the background color of uneditable parts of elements in content items more distinct from the editable parts because previously, it was almost an invisible difference. Now the BG are diagonal stripes of light colors that are unobtrusive but distinctively different from the rest of the UI. Tested using the Stylish extension.